Add Tool type to tools and services
